requirements.txt
six
Faker
pillow
markdown
python-dateutil

django==4.2
django-environ
django-filter
django-allauth==0.55.0
django-resized
python-decouple
django-ckeditor
django-crispy-forms


django-browser-reload
crispy-bootstrap5

django-phonenumber-field[phonenumbers]
djangorestframework
DateTime
django-jazzmin
